requirements.txt
amqp==5.0.9
asgiref==3.4.1
bcrypt==3.2.0
billiard==3.6.4.0
boto3==1.20.49
botocore==1.23.49
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
Deprecated==1.2.13
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
django-phonenumber-field==6.0.0
django-redis==5.2.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
install==1.3.5
jmespath==0.10.0
kombu==5.2.3
packaging==21.3
phonenumbers==8.12.42
Pillow==9.0.1
prompt-toolkit==3.0.24
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyotp==2.6.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-http-client==3.3.5
pytz==2021.3
redis==4.1.2
requests==2.27.1
s3transfer==0.5.1
sendgrid==6.9.5
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stripe==2.65.0
twilio==7.5.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
